// libs/site/ui/src/lib/flow-form/flow-node/flow-node-input.d.ts
import { UiNode, UiNodeInputAttributes } from "@ory/kratos-client";
import { FormEvent } from "react";

export type ValueSetter = (value: string | number | boolean | undefined) => Promise<void>;

export type FormDispatcher = (e: MouseEvent | FormEvent) => Promise<void>;

export interface FlowNodeInputProps {
  node: UiNode;
  attributes: UiNodeInputAttributes;
  value: any;
  disabled: boolean;
  dispatchSubmit: FormDispatcher;
  setValue: ValueSetter;
}